MQTT: Only push messages to MQTT broker if unchanged
Most of these parameters don't change within a minute so there's no point
in pushing them to the MQTT server every few seconds and spam the network/mqtt server.

This patch compares the current packet with the previous one to decide
if it needs to send an updated MQTT message. The time changes every
minute which makes it a convenient keep-alive message notification to verify
that the device didn't drop off the network.